refactor: centralize platform gate in getMessagingProvider instead of 21 call sites
Move assertPlatformEnabled into the registry's getMessagingProvider so
every caller — gmail tools, slack tools, generic messaging tools — gets
the check automatically. Reverts the per-file assertPlatformEnabled
calls added in the previous commit.
