+++ assistant/src/config/bundled-skills/messaging/tools/gmail-archive.ts
@@ -5,7 +5,7 @@ import type {
   ToolContext,
   ToolExecutionResult,
 } from "../../../../tools/types.js";
-import { assertPlatformEnabled, err, ok } from "./shared.js";
+import { err, ok } from "./shared.js";
 
 export async function run(
   input: Record<string, unknown>,
@@ -18,7 +18,6 @@ export async function run(
   }
 
   try {
-    assertPlatformEnabled("gmail");
     const provider = getMessagingProvider("gmail");
     return withValidToken(provider.credentialService, async (token) => {
       await modifyMessage(token, messageId, { removeLabelIds: ["INBOX"] });
