+++ assistant/src/config/bundled-skills/messaging/tools/gmail-label.ts
@@ -5,7 +5,7 @@ import type {
   ToolContext,
   ToolExecutionResult,
 } from "../../../../tools/types.js";
-import { assertPlatformEnabled, err, ok } from "./shared.js";
+import { err, ok } from "./shared.js";
 
 export async function run(
   input: Record<string, unknown>,
@@ -20,7 +20,6 @@ export async function run(
   }
 
   try {
-    assertPlatformEnabled("gmail");
     const provider = getMessagingProvider("gmail");
     return withValidToken(provider.credentialService, async (token) => {
       await modifyMessage(token, messageId, { addLabelIds, removeLabelIds });
