+++ assistant/src/config/bundled-skills/messaging/tools/gmail-send-draft.ts
@@ -5,7 +5,7 @@ import type {
   ToolContext,
   ToolExecutionResult,
 } from "../../../../tools/types.js";
-import { assertPlatformEnabled, err, ok } from "./shared.js";
+import { err, ok } from "./shared.js";
 
 export async function run(
   input: Record<string, unknown>,
@@ -15,7 +15,6 @@ export async function run(
   if (!draftId) return err("draft_id is required.");
 
   try {
-    assertPlatformEnabled("gmail");
     const provider = getMessagingProvider("gmail");
     return withValidToken(provider.credentialService, async (token) => {
       const msg = await sendDraft(token, draftId);
