+++ assistant/src/config/bundled-skills/messaging/tools/gmail-sender-digest.ts
@@ -10,7 +10,7 @@ import type {
   ToolExecutionResult,
 } from "../../../../tools/types.js";
 import { storeScanResult } from "./scan-result-store.js";
-import { assertPlatformEnabled, err, ok } from "./shared.js";
+import { err, ok } from "./shared.js";
 
 const MAX_MESSAGES_CAP = 10000;
 const MAX_IDS_PER_SENDER = 5000;
@@ -58,7 +58,6 @@ export async function run(
   const inputPageToken = input.page_token as string | undefined;
 
   try {
-    assertPlatformEnabled("gmail");
     const provider = getMessagingProvider("gmail");
     return withValidToken(provider.credentialService, async (token) => {
       // Pipeline: fire metadata fetches for each page of IDs as they arrive,
