+++ assistant/src/config/bundled-skills/messaging/tools/gmail-summarize-thread.ts
@@ -14,7 +14,7 @@ import type {
   ToolContext,
   ToolExecutionResult,
 } from "../../../../tools/types.js";
-import { assertPlatformEnabled, err, ok } from "./shared.js";
+import { err, ok } from "./shared.js";
 
 function extractHeader(
   headers: Array<{ name: string; value: string }> | undefined,
@@ -66,7 +66,6 @@ export async function run(
   }
 
   try {
-    assertPlatformEnabled("gmail");
     const provider = getMessagingProvider("gmail");
     return withValidToken(provider.credentialService, async (token) => {
       const listResp = await listMessages(token, `thread:${threadId}`, 50);
