+++ assistant/src/config/bundled-skills/messaging/tools/gmail-unsubscribe.ts
@@ -13,12 +13,7 @@ import type {
   ToolContext,
   ToolExecutionResult,
 } from "../../../../tools/types.js";
-import {
-  assertPlatformEnabled,
-  err,
-  ok,
-  pinnedHttpsRequest,
-} from "./shared.js";
+import { err, ok, pinnedHttpsRequest } from "./shared.js";
 
 export async function run(
   input: Record<string, unknown>,
@@ -37,7 +32,6 @@ export async function run(
   }
 
   try {
-    assertPlatformEnabled("gmail");
     const provider = getMessagingProvider("gmail");
     return withValidToken(provider.credentialService, async (token) => {
       const message = await getMessage(token, messageId, "metadata", [
